+++ src/impls/unsafe/mod.rs
@@ -1,5 +1,4 @@
 mod array;
 mod slice_guard;
 
-pub use slice_guard::DisarmError;
 pub use slice_guard::SliceGuard;