+++ src/impls/vector.rs
@@ -1,0 +1,9 @@
+use alloc::vec::Vec;
+
+use crate::{Capacity, SizeHint};
+
+impl<T> Capacity for Vec<T> {
+    fn capacity_hint(&self) -> SizeHint {
+        SizeHint::UNIVERSAL
+    }
+}