+++ src/lib.rs
@@ -24,3 +24,6 @@ pub use traits::*;
 pub mod either {
     pub use either::Either;
 }
+
+/// Re-export of the `SizeHint` type from the `size_hinter` crate.
+pub use size_hinter::SizeHint;