+++ tests/errors/main.rs
@@ -1,8 +1,8 @@
 mod error_tests;
 
-mod capacity_mismatch;
-mod collection_collision;
+mod capacity_error;
 mod collection_error;
+mod collision;
 mod result_collection_error;
 mod tuple_extension_error;
 mod unzip_error;